reduce responsibility
