+++ deno/v0.1/mod.ts
@@ -1,6 +1,6 @@
 export type {
   BroadcasterInterface,
-  MessageInterface,
+  MinimalRouterState,
 } from "./type_flyweight/router.ts";
 
 export { Router } from "./router/router.ts";
