+++ docs/deno/deps.ts
@@ -1,3 +1,3 @@
-export type { MessageInterface } from "../../deno/v0.1/mod.ts";
+export type { MinimalRouterState } from "../../deno/v0.1/mod.ts";
 
-export { push } from "../../dom/v0.1/mod.ts";
+export { Router } from "../../dom/v0.1/mod.ts";
